# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_abandon_plan_api_v1_plans__plan_id__abandon_post": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why you decided not to buy. Recorded in the plan's journal.",
            "title": "Reason"
          }
        },
        "title": "Body_abandon_plan_api_v1_plans__plan_id__abandon_post",
        "type": "object"
      },
      "ChecklistItemRead": {
        "additionalProperties": false,
        "description": "One pre-buy question and the user's answer.",
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null means unanswered, which is not the same as 'no'.",
            "title": "Answer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "key",
          "question"
        ],
        "title": "ChecklistItemRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_TradePlanRead_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TradePlanRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of matching rows, ignoring pagination.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TradePlanRead]",
        "type": "object"
      },
...
      "TradePlanCreate": {
        "additionalProperties": false,
        "description": "Start a plan for a company.\n\nOnly the symbol is required: the guide's flow is to open a plan and work\nthrough the checklist, so demanding a complete plan up front would push users\ninto filling it in somewhere else and pasting the result.",
        "properties": {
          "intended_amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Amount"
          },
          "invalidation_note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Note"
          },
          "profit_target_pct": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Target Pct"
          },
          "stop_loss_pct": {
            "anyOf": [
              {
                "exclusiveMaximum": 100.0,
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Pct"
          },
          "symbol": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "thesis": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "TradePlanCreate",
        "type": "object"
      },
      "TradePlanDetail": {
        "additionalProperties": false,
        "description": "A single plan, including its review journal.\n\nSeparate from :class:`TradePlanRead` so the list endpoint does not load every\nplan's full review history just to render a table.",
        "properties": {
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItemRead"
            },
            "title": "Checklist",
            "type": "array"
          },
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "company_id": {
            "title": "Company Id",
            "type": "integer"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "intended_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Amount"
          },
          "invalidation_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Note"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "position_sizing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionSizingCheck"
              },
              {
                "type": "null"
              }
            ]
          },
          "profit_target_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Target Pct"
          },
          "readiness": {
            "$ref": "#/components/schemas/TradePlanReadiness"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/PlanReviewRead"
            },
            "title": "Reviews",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TradePlanStatus"
          },
          "stop_loss_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Pct"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "thesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company_id",
          "symbol",
          "company_name",
          "status",
          "checklist",
          "created_at",
          "updated_at",
          "readiness"
        ],
        "title": "TradePlanDetail",
        "type": "object"
      },
...
      "TradePlanReviewCreate": {
        "additionalProperties": false,
        "description": "Record a thesis check-in (guide section 6).\n\nThe note is mandatory. A review that records only a timestamp silences the\nTHESIS_REVIEW_DUE alert without the user having actually re-examined\nanything, which would make the alert worse than useless.",
        "properties": {
          "note": {
            "description": "What you re-checked, and whether the original reason to own it still holds.",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Note",
            "type": "string"
          },
          "thesis_still_valid": {
            "default": true,
            "title": "Thesis Still Valid",
            "type": "boolean"
          }
        },
        "required": [
          "note"
        ],
        "title": "TradePlanReviewCreate",
        "type": "object"
      },
      "TradePlanStatus": {
        "description": "Lifecycle of a pre-commitment plan (guide sections 5 and 6).",
        "enum": [
          "draft",
          "ready",
          "executed",
          "abandoned",
          "closed"
        ],
        "title": "TradePlanStatus",
        "type": "string"
      },
      "TradePlanUpdate": {
        "additionalProperties": false,
        "description": "Patch a draft plan.\n\nEvery field is optional, and ``None`` means \"not supplied\" rather than \"clear\nit\". Clearing an exit rule that has already been committed to is exactly the\nin-the-moment decision the guide warns against, so it is not offered as a\nside effect of an edit.",
        "properties": {
          "comfortable_with_drawdown": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comfortable With Drawdown"
          },
          "debt_manageable_vs_peers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt Manageable Vs Peers"
          },
          "intended_amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intended Amount"
          },
          "invalidation_note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Note"
          },
          "position_size_appropriate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Size Appropriate"
          },
          "profit_target_pct": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Target Pct"
          },
          "revenue_and_profit_healthy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue And Profit Healthy"
          },
          "stop_loss_pct": {
            "anyOf": [
              {
                "exclusiveMaximum": 100.0,
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Pct"
          },
          "thesis": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis"
          },
          "understands_business": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Understands Business"
          }
        },
        "title": "TradePlanUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for evaluating Pakistan Stock Exchange (PSX) equities.\n\n**What this API does:** scores a company against a fixed, published fundamentals\nchecklist; reports technical readings as timing context; records the pre-buy\nchecklist, profit target and stop-loss you commit to *before* buying; tracks your\nportfolio against the diversification limits you set for yourself; and tells you\nwhen one of your own rules has been crossed.\n\n**What it does not do:** predict prices, rate stocks, or tell you what to buy or\nsell. There is no endpoint that returns a recommendation, and that is a deliberate\nproduct constraint rather than a missing feature.\n\nCheck `GET /api/v1/meta` for `provider.is_synthetic`. When it is `true`, every\nfinancial figure and price served by this API is **generated demonstration data**\nfor real PSX ticker symbols - useful for exercising the application, useless for\nreal decisions.\n",
    "title": "PSX Invest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "description": "Plans newest first, optionally filtered by status.",
        "operationId": "list_plans_api_v1_plans_get",
        "parameters": [
          {
            "description": "Restrict results to one plan status.",
            "in": "query",
            "name": "plan_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TradePlanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict results to one plan status.",
              "title": "Plan Status"
            }
          },
          {
            "description": "Rows per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Rows per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TradePlanRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List trade plans",
        "tags": [
          "trade plans"
        ]
      },
      "post": {
        "description": "Open a draft plan for a company.\n\nEverything except the symbol is optional at this point - the plan is meant to\nbe worked through, and the response's ``readiness`` field lists exactly what is\nstill missing before it can be committed to.",
        "operationId": "create_plan_api_v1_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradePlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a trade plan",
        "tags": [
          "trade plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "description": "One plan with its readiness assessment, position sizing and review journal.",
        "operationId": "get_plan_api_v1_plans__plan_id__get",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a trade plan",
        "tags": [
          "trade plans"
        ]
      },
      "patch": {
        "description": "Answer checklist questions, or set the thesis and exit rules.\n\nOnly draft plans are editable. Once a plan has been committed to it is a record\nof a decision, and editing it would rewrite the history the review journal\nexists to preserve.",
        "operationId": "update_plan_api_v1_plans__plan_id__patch",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradePlanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a draft plan",
        "tags": [
          "trade plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}/abandon": {
      "post": {
        "description": "Record that you decided not to buy after all.\n\nThe plan is kept, not deleted: a decision not to act, and the reason for it, is\nworth re-reading the next time the same idea comes around.",
        "operationId": "abandon_plan_api_v1_plans__plan_id__abandon_post",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_abandon_plan_api_v1_plans__plan_id__abandon_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abandon a plan",
        "tags": [
          "trade plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}/close": {
      "post": {
        "description": "Mark a plan closed once the position it governed has been exited.",
        "operationId": "close_plan_api_v1_plans__plan_id__close_post",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close an executed plan",
        "tags": [
          "trade plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}/commit": {
      "post": {
        "description": "Move a plan from draft to ready.\n\nRejected unless all five pre-buy questions are answered \"yes\" **and** both a\nprofit target and a stop-loss are set. That is the one hard rule in this\napplication: the guide's instruction is to decide the exit before buying, not\nwhile watching the price move.",
        "operationId": "commit_plan_api_v1_plans__plan_id__commit_post",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "description": "The checklist is incomplete or the exit rules are missing. The response lists every blocking reason."
          }
        },
        "summary": "Commit to a plan",
        "tags": [
          "trade plans"
        ]
      }
    },
    "/api/v1/plans/{plan_id}/reviews": {
      "post": {
        "description": "Record that you re-checked why you own this, and what you concluded.\n\nThis is the guide's \"thesis check-in\": if the reason you bought has changed,\nthat is a far stronger signal than a red candle on a chart. Recording a review\nclears the review-due alert, which is why a substantive note is required.",
        "operationId": "record_review_api_v1_plans__plan_id__reviews_post",
        "parameters": [
          {
            "description": "Trade plan id.",
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "description": "Trade plan id.",
              "minimum": 1,
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradePlanReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradePlanDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Plan not found."
          },
          "409": {
            "description": "The plan is not in a state that allows this."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a thesis check-in",
        "tags": [
          "trade plans"
        ]
      }
    },
...
  }
}